requirements.txt
DrissionPage~=4.0.5.6
elasticsearch~=8.14.0
json5~=0.9.25
pypandoc~=1.13
python_docx~=1.1.2
qwen_agent~=0.0.6
sentence_transformers~=3.0.1
tqdm~=4.66.4
